+++ src/main/kotlin/com/eitanliu/dart/mappable/settings/SettingLayout.kt
@@ -35,6 +35,15 @@ class SettingLayout(private val settings: Settings) : UnnamedConfigurable {
 
             rowComment("Configure dart data model files suffix.")
         }
+        row {
+            checkBox(
+                "auto run build_runner"
+            ).bindSelected(
+                graph.autoBuildRunner
+            ).applyToComponent {
+                toolTipText = "auto run 'flutter pub run build_runner build --delete-conflicting-outputs'"
+            }
+        }
         buttonGroup(graph.implement, "Implement") {
             row {
                 checkBox(
@@ -58,6 +67,15 @@ class SettingLayout(private val settings: Settings) : UnnamedConfigurable {
             rowsRange {
                 buildMappable()
             }.visibleIf(mappablePredicate)
+            val freezedPredicate = graph.implement.selected(Implements.FREEZED)
+            row {
+                radioButton("freezed", Implements.FREEZED)
+                    .bindSelected(graph.implement, Implements.FREEZED)
+            }
+            rowsRange {
+                buildFreezed()
+            }.visibleIf(freezedPredicate)
+        }.bind(::implement)
         }
 
     }
@@ -102,13 +120,13 @@ class SettingLayout(private val settings: Settings) : UnnamedConfigurable {
                         bindText(graph.mappableToJson)
                     }.horizontalAlign(HorizontalAlign.FILL)
                 }.layout(RowLayout.LABEL_ALIGNED)
-                // row {
-                //     checkBox("copyWith")
-                //         .bindSelected(graph.enableCopyWith)
-                //     textField().apply {
-                //         bindText(graph.mappableCopyWith)
-                //     }.horizontalAlign(HorizontalAlign.FILL)
-                // }.layout(RowLayout.LABEL_ALIGNED)
+                row {
+                    checkBox("copyWith")
+                        .bindSelected(graph.enableCopyWith)
+                    textField().apply {
+                        bindText(graph.mappableCopyWith)
+                    }.horizontalAlign(HorizontalAlign.FILL)
+                }.layout(RowLayout.LABEL_ALIGNED)
             }.visibleIf(customPredicate)
         }
     }
@@ -147,6 +165,15 @@ class SettingLayout(private val settings: Settings) : UnnamedConfigurable {
         }
     }
 
+    private fun Panel.buildFreezed() {
+        indent {
+            row {
+                checkBox("fromJson/toJson")
+                    .bindSelected(graph.freezedEnableJson)
+            }
+        }
+    }
+
     private fun Panel.rowPanel(
         title: String? = null,
         indent: Boolean = true,
@@ -167,6 +194,7 @@ class SettingLayout(private val settings: Settings) : UnnamedConfigurable {
     override fun isModified(): Boolean {
         return settings.graph.modelSuffix.value != graph.modelSuffix.value
                 || settings.graph.implement.value != graph.implement.value
+                || settings.graph.autoBuildRunner.value != graph.autoBuildRunner.value
                 || settings.graph.enableJsonReflectable.value != graph.enableJsonReflectable.value
                 || settings.graph.enableMixin.value != graph.enableMixin.value
                 || settings.graph.enableFromJson.value != graph.enableFromJson.value
@@ -179,11 +207,13 @@ class SettingLayout(private val settings: Settings) : UnnamedConfigurable {
                 || settings.graph.mappableFromMap.value != graph.mappableFromMap.value
                 || settings.graph.mappableToMap.value != graph.mappableToMap.value
                 || settings.graph.mappableCopyWith.value != graph.mappableCopyWith.value
+                || settings.graph.freezedEnableJson.value != graph.freezedEnableJson.value
     }
 
     override fun apply() {
         settings.graph.modelSuffix.value = graph.modelSuffix.value
         settings.graph.implement.value = graph.implement.value
+        settings.graph.autoBuildRunner.value = graph.autoBuildRunner.value
         settings.graph.enableJsonReflectable.value = graph.enableJsonReflectable.value
         settings.graph.enableMixin.value = graph.enableMixin.value
         settings.graph.enableFromJson.value = graph.enableFromJson.value
@@ -196,11 +226,13 @@ class SettingLayout(private val settings: Settings) : UnnamedConfigurable {
         settings.graph.mappableFromMap.value = graph.mappableFromMap.value
         settings.graph.mappableToMap.value = graph.mappableToMap.value
         settings.graph.mappableCopyWith.value = graph.mappableCopyWith.value
+        settings.graph.freezedEnableJson.value = graph.freezedEnableJson.value
     }
 
     override fun reset() {
         graph.modelSuffix.value = settings.graph.modelSuffix.value
         graph.implement.value = settings.graph.implement.value
+        graph.autoBuildRunner.value = settings.graph.autoBuildRunner.value
         graph.enableJsonReflectable.value = settings.graph.enableJsonReflectable.value
         graph.enableMixin.value = settings.graph.enableMixin.value
         graph.enableFromJson.value = settings.graph.enableFromJson.value
@@ -213,6 +245,7 @@ class SettingLayout(private val settings: Settings) : UnnamedConfigurable {
         graph.mappableFromMap.value = settings.graph.mappableFromMap.value
         graph.mappableToMap.value = settings.graph.mappableToMap.value
         graph.mappableCopyWith.value = settings.graph.mappableCopyWith.value
+        graph.freezedEnableJson.value = settings.graph.freezedEnableJson.value
     }
 
     class Graph(data: SettingLayout) {
@@ -220,6 +253,7 @@ class SettingLayout(private val settings: Settings) : UnnamedConfigurable {
 
         val modelSuffix = propertyGraph.propertyOf(data.settings.modelSuffix)
         val implement = propertyGraph.propertyOf(data.settings.implement)
+        val autoBuildRunner = propertyGraph.propertyOf(data.settings.autoBuildRunner)
         val enableJsonReflectable = propertyGraph.propertyOf(data.settings.enableJsonReflectable)
         val enableMixin = propertyGraph.propertyOf(data.settings.enableMixin)
         val enableFromJson = propertyGraph.propertyOf(data.settings.enableFromJson)
@@ -232,6 +266,7 @@ class SettingLayout(private val settings: Settings) : UnnamedConfigurable {
         val mappableFromMap = propertyGraph.propertyOf(data.settings.mappableFromMap)
         val mappableToMap = propertyGraph.propertyOf(data.settings.mappableToMap)
         val mappableCopyWith = propertyGraph.propertyOf(data.settings.mappableCopyWith)
+        val freezedEnableJson = propertyGraph.propertyOf(data.settings.freezedEnableJson)
 
     }
 }